SMB-87: Corrections de warnings
Principalement des comparaisons d'entiers signés et non-signés.

Aussi corrigé un rouspétage de Sonar sur un constructeur.

+++ src/modeles/scriptorium/Scribe.cpp
@@ -31,7 +31,7 @@ namespace sambada {
 	{
 		Archiviste::initialise(nomsFichiers, chaineRetourLigne, charDelimMots, precisionFlots);
 
-		for(int i(0); i < noms.size(); ++i)
+		for(size_t i(0); i < noms.size(); ++i)
 		{
 			flots.push_back(factory.creeFlotSortie(noms[i]));
 			flots[i]->precision(precision);
